py-service/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pillow
sqlalchemy
psycopg2-binary
python-dotenv
